Refactor: lib/usage 내부 모듈 계층 구조화 및 전역 참조 경로 최적화
1. 기능 및 가독성 기반 디렉토리 구조 고도화

모듈 책임 분리: lib/usage 루트에 산재해 있던 10개의 공용 모듈을 도메인 및 역할에 따라 messaging/, api/, hooks/ 계층으로 분류 배치하여 유기적인 관리 체계를 구축했습니다.

messaging/: 에러 코드 마스킹, 메시지 및 빈 상태 배너 관리 (dashboard-messages, team-bff-fetch-errors 등)

api/: 게이트웨이 및 BFF 호출, 베이스 URL 구성 (fetch-usage, usage-gateway-bff-proxy 등)

hooks/: 필터 상태 보존, 스토리지 바인딩 및 데이터 커스텀 훅 (use-filter-storage, usage-log-tab-storage 등)

(단, 전체 파일에서 범용적으로 쓰이는 types, format, kst-dates 등의 파일은 루트 수준에 유지하여 불필요한 변경 레이어를 줄였습니다.)

2. 전역 Import 경로 일괄 정렬 및 동기화

소비처 참조 최신화: 모듈 이동에 따라 영향을 받는 usage-web 내 대시보드 컴포넌트 전체(usage-dashboard, team-dashboard 등), 사용량 로그 패널, 필터 바, Next.js API 엔드포인트 라우트의 상단 명세를 신규 계층 경로에 맞춰 일괄 갱신했습니다.

3. 안정성 및 잔존 레거시 검증

정적 분석: 리팩토링 완료 후 프로젝트 내에 기존 구버전 경로(@/lib/usage/fetch-usage 등)를 바라보는 코드가 남아있지 않음을 grep 전수 조사를 통해 검증했습니다.

클라이언트 테스트: 영향권에 있는 messaging, hooks, api route 관련 Vitest 유닛 테스트를 수행하여 전체 25개 케이스가 모두 정상 통과(25 passed)함을 확인했습니다.

범위 격리: 웹 영역에 국한된 파일 이동이므로 Java 백엔드 및 인프라 코드는 일절 변경 사항이 없습니다.
