+++ services/usage-service/web/src/app/api/usage/[[...path]]/route.test.ts
@@ -1,10 +1,10 @@
-import { afterEach, describe, expect, it, vi } from "vitest"
+﻿import { afterEach, describe, expect, it, vi } from "vitest"
 
 import {
   USAGE_PROXY_MESSAGE_404,
   USAGE_PROXY_MESSAGE_500,
   USAGE_PROXY_MESSAGE_502,
-} from "@/lib/usage/usage-gateway-bff-proxy"
+} from "@/lib/usage/api/usage-gateway-bff-proxy"
 
 import { GET } from "./route"
