+++ services/usage-service/web/src/app/api/usage/team/bff/[[...path]]/route.ts
@@ -1,8 +1,8 @@
-import {
+﻿import {
   encodeUsagePathSegments,
   proxyUsageToGateway,
   usageProxyJsonError,
-} from "@/lib/usage/usage-gateway-bff-proxy"
+} from "@/lib/usage/api/usage-gateway-bff-proxy"
 
 type RouteContext = { params: Promise<{ path?: string[] }> }
